AI_Projects/Resume_Critiquer/pyproject.toml
[project]
name = "python-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.102.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.0",
]

[tool.uv.workspace]
members = [
    "Image_Classifier",
]
